pyproject.toml
[project]
name = "telethon-core"
version = "0.1.0"
description = "A modular Telethon based bot core."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "python-dotenv==1.0.1",
    "telethon==1.36.0",
]
